server/pdf_chat_project/requirements.txt
django
djangorestframework
django-cors-headers
PyPDF2
langchain
google-generativeai
faiss-cpu
python-dotenv